Scaling, Interface, and Comments
Added scaling parameters to interface, showing the final result after
scaling along the bottom of the form.  Also added public methods such
that the form doesn't need to be initialised.  Added comments to the top
of each class to indicate the public variables and a brief description
of the operation of each method within the class.  Deleted LabView
files.